Generalize volume mapping in docker script, #17

+++ tools/start-docker.sh
@@ -10,6 +10,7 @@
 set -x
 
 TYPO3VERSION=7
+MAIN_DIR=$( cd "$( dirname "${BASH_SOURCE[0]}" )"/.. && pwd )
 
 # Stop already running containers
 if docker ps | grep -q typo3-web ; then
@@ -21,10 +22,16 @@ fi
 
 if [[ "$1" = "--new" ]]; then
     if docker ps -a | grep -q typo3-db ; then
-        docker rm typo3-db
+        read -p "Delete the existing typo3-db image? " -n 1 -r
+        if [[ $REPLY =~ ^[Yy]$ ]]; then
+            docker rm typo3-db
+        fi
     fi
     if docker ps -a | grep -q typo3-web ; then
-        docker rm typo3-web
+        read -p "Delete the existing typo3-web image? " -n 1 -r
+        if [[ $REPLY =~ ^[Yy]$ ]]; then
+            docker rm typo3-web
+        fi
     fi
     echo "Renew docker containers (start from scratch again)"
     # Run container with database
@@ -40,7 +47,7 @@ if [[ "$1" = "--new" ]]; then
     docker run -d --name typo3-web \
         --link typo3-db:db \
         -p 80:80 \
-        -v `pwd`:/var/www/html/typo3conf/ext/publist4ubma2 \
+        -v "$MAIN_DIR":/var/www/html/typo3conf/ext/publist4ubma2 \
       martinhelmich/typo3:"$TYPO3VERSION"
 else
     if ! docker ps -a | grep -q typo3-db ; then
